+++ jkorm/src/test/kotlin/com/tang/jkorm/sql/SqlProviderTest.kt
@@ -47,6 +47,4 @@ interface SqlProviderTest {
 
     fun paginateOrderBy()
 
-    fun paginateLimit()
-
 }